# server/src/api_documentation.json
{
    "openapi": "3.0.4",
    "info": {
        "title": "Vibezone API",
        "description": "Ta dokumentacja opisuje API dla aplikacji Vibezone.",
        "version": "0.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
...
    "components": {
        "schemas": {
            "Profile": {
                "description": "Profil użytkownika zawierający informacje o nim.",
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "description": "Unikalny identyfikator użytkownika",
                        "format": "objectId"
                    },
                    "username": {
                        "type": "string",
                        "description": "Unikalna nazwa użytkownika"
                    },
                    "displayName": {
                        "type": "string",
                        "description": "Nazwa wyświetlana użytkownika."
                    },
                    "profilePictureUri": {
                        "type": "string",
                        "description": "URI dla zdjęcia profilowego.",
                        "format": "uri"
                    },
                    "aboutDesc": {
                        "type": "string",
                        "description": "Opis/życiorys użytkownika."
                    },
                    "createdAt": {
                        "type": "string",
                        "description": "Data kiedy profil został utworzony.",
                        "format": "date-time"
                    }
                },
                "required": ["_id", "username", "createdAt"]
            },
            "ProfileLite": {
                "description": "Mniejsza wersja schematu Profile zawarty w schematach postów i komentarzy.",
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "description": "Unikalny identyfikator użytkownika - autora posta",
                        "format": "objectId"
                    },
                    "username": {
                        "type": "string",
                        "description": "Unikalna nazwa użytkownika"
                    },
                    "displayName": {
                        "type": "string",
                        "description": "Opcjonalna nazwa wyświetlana użytkownika."
                    },
                    "profilePictureUri": {
                        "type": "string",
                        "description": "Opcjonalny URI dla zdjęcia profilowego.",
                        "format": "uri"
                    }
                },
                "required": ["_id", "username"]
            },
            "Post": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "description": "Unikalny identyfikator posta",
                        "format": "objectId"
                    },
                    "author": {
                        "$ref": "#/components/schemas/ProfileLite"
                    },
                    "content": {
                        "type": "string",
                        "description": "Treść posta."
                    },
                    "imageUrl": {
                        "type": "string",
                        "description": "Zdjęcie posta.",
                        "format": "uri"
                    },
                    "createdAt": {
                        "type": "string",
                        "description": "Data kiedy post został utworzony.",
                        "format": "date-time"
                    },
                    "likeCount": {
                        "type": "number",
                        "description": "Ilość polubień posta."
                    },
                    "commentCount": {
                        "type": "number",
                        "description": "Ilość komentarzy pod postem."
                    },
                    "isLikedByUser": {
                        "type": "boolean",
                        "description": "Czy ten post jest polubiony przez aktualnie zautoryzowanego użytkownika. Zawsze `false` jeżeli niezalogowany."
                    }
                },
                "required": [
                    "_id",
                    "author",
                    "content",
                    "createdAt",
                    "likeCount",
                    "commentCount"
                ]
            },
...
            "UnauthorizedError": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Unauthorized"
                    }
                }
            },
            "ForbiddenError": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Forbidden"
                    }
                }
            },
            "NotFoundError": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Not found"
                    }
                }
            },
            "ValidationError": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "description": "Błąd wiadomości (co poszło nie tak).",
                        "example": "Display name must not be longer than 32 characters."
                    },
                    "item": {
                        "type": "string",
                        "description": "Pole formularza lub wyboru w którym doszło do błędu walidacji.",
                        "example": "displayName"
                    }
                },
                "required": ["error", "item"]
            },
            "MalformedIdError": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Malformed id"
                    }
                }
            }
        },
        "securitySchemes": {
            "sessionCookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "session"
            }
        }
    },
    "paths": {
...
        "/profile/{id}": {
            "get": {
                "tags": ["profile"],
                "summary": "Pobranie profilu użytkownika z `id`.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string",
                            "format": "objectId"
                        },
                        "required": true,
                        "description": "Identyfikator profilu użytkownika który ma zostać pobrany."
                    }
                ],
                "security": [],
                "responses": {
                    "200": {
                        "description": "Pomyślnie pobrano profil użytkownika.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Profile"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Nieprawidłowa forma `id`.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MalformedIdError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nie znaleziono profilu o takim id.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/profile": {
            "get": {
                "tags": ["profile"],
                "summary": "Pobranie profilu zalogowanego użytkownika.",
                "security": [
                    {
                        "sessionCookie": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pomyślnie pobrano profil zalogowanego użytkownika.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Profile"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "role": {
                                                    "type": "string",
                                                    "description": "Rola użytkownika - `user` albo `moderator`.",
                                                    "example": "moderator"
                                                }
                                            },
                                            "required": ["role"]
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Użytkownik nie jest zalogowany.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/profile/update": {
            "patch": {
                "tags": ["profile"],
                "summary": "Aktualizacja informacji profilu użytkownika.",
                "security": [
                    {
                        "sessionCookie": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "description": "Wszystkie właściwości tutaj są nullable i opcjonalne, przy czym: - wartość tekstowa: informacja zostanie zaktualizowana; - wartość typu `null` w miejscu właściwości: informacja zostanie usunięta z profilu; - brak właściwości w zapytaniu: informacja zostanie bez zmian w profilu.\n",
                                "properties": {
                                    "displayName": {
                                        "type": "string",
                                        "description": "Nazwa wyświetlana użytkownika.",
                                        "nullable": true
                                    },
                                    "aboutDesc": {
                                        "type": "string",
                                        "description": "Opis/życiorys użytkownika.",
                                        "nullable": true
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pomyślnie zaktualizowano profil użytkownika. Zwracany jest obiekt profilu po edycji.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Profile"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Błąd walidacji danych.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Użytkownik nie jest zalogowany.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/profile/picture": {
            "post": {
                "tags": ["profile"],
                "summary": "Aktualizacja zdjęcia profilowego.",
                "security": [
                    {
                        "sessionCookie": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "avatar": {
                                        "type": "string",
                                        "format": "binary",
                                        "description": "Plik wysyłanego zdjęcia profilowego."
                                    }
                                },
                                "required": ["avatar"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pomyślnie zaktualizowano zdjęcie profilowe użytkownika. Zwracany jest obiekt profilu po edycji.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Profile"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Odrzucono zdjęcie. Możliwe przyczyny to: zły rodzaj pliku lub za duży rozmiar (>10MB)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No proper file uploaded."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Użytkownik nie jest zalogowany.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["profile"],
                "summary": "Usuwanie zdjęcia profilowego.",
                "security": [
                    {
                        "sessionCookie": []
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Pomyślnie usunięto zdjęcie profilowe użytkownika. Zwracany jest obiekt profilu po edycji."
                    },
                    "401": {
                        "description": "Użytkownik nie jest zalogowany.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/profile/{id}/posts": {
            "get": {
                "tags": ["profile"],
                "summary": "Pobieranie wszystkich postów użytkownika",
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string",
                            "format": "objectId"
                        },
                        "required": true,
                        "description": "Identyfikator użytkownika które posty mają zostać zwrócone."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pomyślnie pobrano listę wszystkich postów użytkownika.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Post"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Nieprawidłowa forma `id`.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MalformedIdError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nie znaleziono profilu o takim id.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/profile/{id}/ban": {
            "post": {
                "tags": ["profile"],
                "summary": "Zbanowanie użytkownika i usunięcie jego profilu, postów i komentarzy.",
                "security": [
                    {
                        "sessionCookie": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string",
                            "format": "objectId"
                        },
                        "required": true,
                        "description": "Identyfikator użytkownika który ma zostać zbanowany."
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Pomyślnie zbanowano użytkownika i usunięto jego danych."
                    },
                    "400": {
                        "description": "Nieprawidłowa forma `id`.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MalformedIdError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Użytkownik nie ma uprawnień do banowania użytkowników (nie jest moderatorem).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nie znaleziono profilu o takim id.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}